Update liquidity type in Tick and CurrentLiquidity. Migration to fix overflowing liquidity ticks and reserves.

+++ pallets/admin-utils/src/tests/mock.rs
@@ -333,7 +333,7 @@ parameter_types! {
     pub const SwapProtocolId: PalletId = PalletId(*b"ten/swap");
     pub const SwapMaxFeeRate: u16 = 10000; // 15.26%
     pub const SwapMaxPositions: u32 = 100;
-    pub const SwapMinimumLiquidity: u64 = 1_000;
+    pub const SwapMinimumLiquidity: u128 = 1_000;
     pub const SwapMinimumReserve: NonZeroU64 = NonZeroU64::new(1_000_000).unwrap();
 }
