+++ pallets/subtensor/src/tests/mock.rs
@@ -476,7 +476,7 @@ parameter_types! {
     pub const SwapProtocolId: PalletId = PalletId(*b"ten/swap");
     pub const SwapMaxFeeRate: u16 = 10000; // 15.26%
     pub const SwapMaxPositions: u32 = 100;
-    pub const SwapMinimumLiquidity: u64 = 1_000;
+    pub const SwapMinimumLiquidity: u128 = 1_000;
     pub const SwapMinimumReserve: NonZeroU64 = NonZeroU64::new(100).unwrap();
 }
